Move forwardEvent to Events.ts
